fixed Entities, Dtos, Repositories, added Mapper.

+++ src/main/java/hay/java/repository/ChapterItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChapterItemRepository extends JpaRepository<ChapterItemEntity, Integer> {
+public interface ChapterItemRepository extends JpaRepository<ChapterItemEntity, Long> {
 }
