+++ src/main/java/hay/java/repository/ChapterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChapterRepository extends JpaRepository<ChapterEntity,Integer> {
+public interface ChapterRepository extends JpaRepository<ChapterEntity,Long> {
 }
